fix: pipeline migration name
